更新IPTV源列表的默认值和配置，完善播放器缓冲设置

+++ tv/src/main/java/top/yogiczy/mytv/tv/ui/screensold/videoplayer/player/IjkVideoPlayer.kt
@@ -74,7 +74,7 @@ class IjkVideoPlayer(
             setOption(IjkMediaPlayer.OPT_CATEGORY_FORMAT, "buffer_size", 1316)
             setOption(IjkMediaPlayer.OPT_CATEGORY_FORMAT, "infbuf", 1)  // 无限读
             setOption(IjkMediaPlayer.OPT_CATEGORY_FORMAT, "flush_packets", 1L)
-            setOption(IjkMediaPlayer.OPT_CATEGORY_PLAYER, "min-frames", Configs.videoPlayerBufferTime.toLong() * 0.03L);
+            setOption(IjkMediaPlayer.OPT_CATEGORY_PLAYER, "min-frames", (Configs.videoPlayerBufferTime.toLong() * 0.03).toLong());
 
             //  关闭播放器缓冲，这个必须关闭，否则会出现播放一段时间后，一直卡住，控制台打印 FFP_MSG_BUFFERING_START
             setOption(IjkMediaPlayer.OPT_CATEGORY_PLAYER, "packet-buffering", 0)
